pyproject.toml
[project]
name = "cv"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Adrian Berges", email = "aberges@outlook.com"},
]
dependencies = [
    "omegaconf>=2.3.0",
    "python-dateutil>=2.9.0.post0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
